feat:  enhance AI chat sidebar panel - integrated Gemini API calls for producton - add option to stream/no-stream responses -  modularized code for better readability - removed bugging features (maybe in future)
