+++ package.json
@@ -12,6 +12,7 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.10.0",
+    "@google/genai": "^1.24.0",
     "@hookform/resolvers": "^5.2.2",
     "@monaco-editor/react": "^4.7.0",
     "@prisma/client": "^6.16.1",
@@ -50,6 +51,7 @@
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
     "lucide-react": "^0.544.0",
+    "moment": "^2.30.1",
     "monaco-editor": "^0.53.0",
     "monacopilot": "^1.2.7",
     "next": "15.5.3",
